docs: Update instructions for creating Watson API keys
I've taken a new set of screenshots for the end-to-end process of
creating an IBM Cloud account, and creating service instances for Watson
Assistant and Visual Recognition.

I updated the grid with costs to reflect the policy changes for Visual
Recognition, and added another worked costs example to the list.

I also added links to the two programmes I'm aware of that are available
to schools to request free usage of Watson APIs.

Closes: IBM/taxinomitis#37

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>