docs: Simplified version of Make Me Happy
I'm running an event tomorrow where we will have a very short time
period to run activities with a large group of kids.

I've simplified Make Me Happy for this event, where kids aren't
asked to log in, and don't create the functional rules-based version
of the character first - instead going straight to the ML-based
version.

I think this should save 10 minutes or so when running the class.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>