docs: Simplified version of Smart Classroom
I'm running an event tomorrow where we will have a very short time
period to run activities with a large group of kids.

I've simplified Smart Classroom for this event, where kids aren't
asked to log in, and don't try the functional rules-based version
of the assistant first - instead going straight to the ML-based
version.

I think this should save 10 minutes or so when running the class.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>